Fixed: Circle.java корректные формулы, Laptop.java нифицырованы поле price, Point.java иннициализировано поле Y, Product.java добавлен пробел

+++ src/main/java/practice_2v/Laptop.java
@@ -4,7 +4,7 @@ public class Laptop {
     String brand;
     double price;
 
-    Laptop(String someBrand, int somePrice){
+    Laptop(String someBrand, double somePrice){
         this.brand = someBrand;
         this.price = somePrice;
